Order rows in more places!
Order is now a option when creating new object.

Set the order of a has_many association.
has_many('articles, array('order' => 'title ASC'))

This leads to enhanced first() and last() methods.

By default first() uses id ASC and last() uses id DESC.

If order is title ASC.

first() uses title ASC and last() uses title DESC.

This is smart because the old functionality was logical about DESC and
ASC. Now it is based on business logic. If you ordering the items in
reverse order than the first item is first item in the list. Not the
last item in the database. :)